+++ ergo_std/ChangeLog.md
@@ -104,6 +104,7 @@
 * Improve `std:fs:copy` to allow non-existent `from` paths if making a symbolic
   link.
 * Rename `std:ergo-remote` to `std:load-remote`.
+* Remove `std:variable` since the builtin `id` does the same thing.
 
 ### Migration
 * Remove the use of `std:by-content`. If it was being used to compare nested
@@ -149,6 +150,7 @@
   ...` and change any use of `self` in the function body to not pass `self` as
   the first argument.
 * Change `std:ergo-remote` to `std:load-remote`.
+* Replace `std:variable ~depends=DEPS VALUE` with `id ~set=DEPS VALUE`.
 
 ## 1.0.0-rc.3  -- 2021-12-14
 * No changes.